+++ src/item/accessible/api.ts
@@ -1,17 +1,13 @@
-import { PackedItem } from '@graasp/sdk';
+import { PackedItem, Paginated, Pagination } from '@graasp/sdk';
 
 import { verifyAuthentication } from '../../api/axios.js';
-import {
-  Paginated,
-  PaginationParams,
-  PartialQueryConfigForApi,
-} from '../../types.js';
+import { PartialQueryConfigForApi } from '../../types.js';
 import { buildGetAccessibleItems } from '../routes.js';
 import { ItemSearchParams } from '../types.js';
 
 export const getAccessibleItems = async (
   params: ItemSearchParams,
-  pagination: PaginationParams,
+  pagination: Partial<Pagination>,
   { API_HOST, axios }: PartialQueryConfigForApi,
 ) =>
   verifyAuthentication(() =>