+++ src/item/routes.ts
@@ -1,7 +1,6 @@
-import { UUID } from '@graasp/sdk';
+import { Pagination, UUID } from '@graasp/sdk';
 
 import { DEFAULT_THUMBNAIL_SIZE } from '../config/constants.js';
-import { PaginationParams } from '../types.js';
 import { ItemChildrenParams, ItemSearchParams } from './types.js';
 
 export const ITEMS_ROUTE = 'items';
@@ -10,8 +9,15 @@ export const GET_RECYCLED_ITEMS_DATA_ROUTE = `${ITEMS_ROUTE}/recycled`;
 export const SHARED_ITEM_WITH_ROUTE = `${ITEMS_ROUTE}/shared-with`;
 
 export const buildGetAccessibleItems = (
-  { creatorId, name, ordering, sortBy, permissions, types }: ItemSearchParams,
-  { page, pageSize }: PaginationParams,
+  {
+    creatorId,
+    ordering,
+    sortBy,
+    permissions,
+    types,
+    keywords,
+  }: ItemSearchParams,
+  { page, pageSize }: Partial<Pagination>,
 ) => {
   const searchParams = new URLSearchParams();
   // pagination params
@@ -23,15 +29,15 @@ export const buildGetAccessibleItems = (
   if (creatorId) {
     searchParams.set('creatorId', creatorId);
   }
-  if (name) {
-    searchParams.set('name', name);
-  }
   if (sortBy) {
     searchParams.set('sortBy', sortBy);
   }
   if (ordering) {
     searchParams.set('ordering', ordering);
   }
+  keywords?.split(' ')?.forEach((k) => {
+    searchParams.append('keywords', k);
+  });
   if (permissions && permissions.length) {
     permissions.forEach((p) => searchParams.append('permissions', p));
   }
@@ -72,7 +78,7 @@ export const buildDeleteItemsRoute = (ids: UUID[]) =>
 
 export const buildGetChildrenRoute = (
   id: UUID,
-  { ordered, types }: ItemChildrenParams,
+  { ordered, types, keywords }: ItemChildrenParams,
 ) => {
   const route = `${ITEMS_ROUTE}/${id}/children`;
   const search = new URLSearchParams();
@@ -82,6 +88,9 @@ export const buildGetChildrenRoute = (
   if (types && types.length) {
     types.forEach((t) => search.append('types', t));
   }
+  keywords?.split(' ')?.forEach((k) => {
+    search.append('keywords', k);
+  });
   if (search.toString()) {
     return `${route}?${search}`;
   }